requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
doc_scanner @ git+https://github.com/M-krosoft/doc-scanner.git@410950fdb7321d1f4850a5ad155a5ee87ab206b5
Flask==3.0.3
gunicorn==20.1.0
idna==3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
imageio==2.35.1
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
SQLAlchemy==2.0.35
tifffile==2024.9.20
urllib3==2.2.3
typing_extensions==4.12.2
Werkzeug==3.0.4
